// src/features/resume/components/Resume/types.ts
interface IResumeFormValues {
  position: string;
  category: string;
  skills: string[];
  workExperience: number;
  experienceLevel: string;
  salaryExpectations: number;
  country: string;
  city: string;
  relocation: boolean;
  englishLevel: string;
  summary: string;
  employment: string[];
  dontConsider: string[];
  isPublished: boolean;
}

export { IResumeFormValues };